+++ CHANGELOG.md
@@ -8,6 +8,8 @@
  - Default video container setting
  - Batch input support
  - Full fallback on TRT upscaling, so the engine has the greatest chance of building correctly.
+ - Prores codec
+ - Additional Animation and Realistic Models
 ### Changed
  - Adjusted dynamic scale.
  - Transcodes audio by default, to fix potential audio copying issues.